requirements.txt
streamlit
pandas
plotly
gspread
google-auth
google-generativeai
mistralai
requests
openai